health-first-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
pymongo==4.6.0
bcrypt==4.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
phonenumbers==8.13.26
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
motor==3.3.2
aiosqlite==0.19.0
pytz==2023.3
python-dateutil==2.8.2 